Initial Commit:
Basic functionality is available. The programm checks if there are four
markers and if the parity is correct. It prints the first four markers
out. If there are less than four there is another return code than zero
and there will be markers with just zero values given out. Same for more
than four makers. If the parity is wrong, there is also an exit code not
zero.